Split rdata::rfc1035 into private sub-modules.
